pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "alfred"
version = "0.1.0"
description = "Alfred — unified vault operations suite"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.27",
    "python-frontmatter>=1.1",
    "structlog>=24.0",
    "watchdog>=4.0",
]

[project.optional-dependencies]
surveyor = [
    "pymilvus>=2.4",
    "scikit-learn>=1.4",
    "leidenalg>=0.10",
    "igraph>=0.11",
    "numpy>=1.26",
    "openai>=1.30",
]
all = ["alfred[surveyor]"]

[project.scripts]
alfred = "alfred.cli:main"
